+++ packages/knowledge-schema/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./enums.js";
 export * from "./metadata.js";
+export * from "./migration.js";
 export * from "./objects.js";
 export * from "./parse.js";
 export * from "./primitives.js";